+++ src/senaite/storage/archetypes/field.py
@@ -1,0 +1,36 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of SENAITE.STORAGE.
+#
+# SENAITE.STORAGE is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by the Free
+# Software Foundation, version 2.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+# details.
+#
+# You should have received a copy of the GNU General Public License along with
+# this program; if not, write to the Free Software Foundation, Inc., 51
+# Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+#
+# Copyright 2019-2026 by it's authors.
+# Some rights reserved, see README and LICENSE.
+
+from archetypes.schemaextender.field import ExtensionField as ATExtensionField
+from Products.Archetypes.atapi import DateTimeField
+
+
+class ExtensionField(ATExtensionField):
+    """Mix-in class to make Archetypes fields not depend on generated accessors
+    and mutators, and use AnnotationStorage by default
+    """
+
+    def __init__(self, *args, **kwargs):
+        super(ExtensionField, self).__init__(*args, **kwargs)
+
+
+class ExtDateTimeField(ExtensionField, DateTimeField):
+    """Field extender of DateTimeField
+    """