+++ src/senaite/storage/browser/container/store_samples.py
@@ -23,10 +23,13 @@ import json
 from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from senaite.storage import api as _api
 from senaite.storage import logger
 from senaite.storage import senaiteMessageFactory as _s
 from senaite.storage.browser import BaseView
 from senaite.storage.catalog import STORAGE_CATALOG
+from DateTime import DateTime
+
 
 DISPLAY_TEMPLATE = "<a href='${url}' _target='blank'>${get_full_title}</a>"
 
@@ -68,6 +71,7 @@ class StoreSamplesView(BaseView):
             # extract relevant data
             container_mapping = form.get("sample_container", {})
             container_position_mapping = form.get("sample_container_position", {})
+            retention_mapping = form.get("sample_retention_period", {})
 
             for sample in samples:
                 sample_uid = api.get_uid(sample)
@@ -86,6 +90,14 @@ class StoreSamplesView(BaseView):
                 if stored:
                     stored = container.get_object_at(position[0], position[1])
                     stored_samples.append(stored)
+                    # Compute and store the expiry date
+                    retention_days = retention_mapping.get(sample_uid)
+                    retention_days = api.to_int(retention_days, default=-1)
+                    if retention_days >= 0:
+                        expiry = DateTime() + retention_days
+                        sample.setStorageExpiryDate(expiry)
+                    else:
+                        sample.setStorageExpiryDate(None)
 
             message = _s("Stored {} samples: {}".format(
                 len(stored_samples), ", ".join(
@@ -104,6 +116,7 @@ class StoreSamplesView(BaseView):
         """
         for obj in self.get_objects_from_request():
             obj = api.get_object(obj)
+            retention = _api.get_default_retention_period(obj)
             yield {
                 "obj": obj,
                 "id": api.get_id(obj),
@@ -111,7 +124,8 @@ class StoreSamplesView(BaseView):
                 "title": api.get_title(obj),
                 "path": api.get_path(obj),
                 "url": api.get_url(obj),
-                "sample_type": api.get_title(obj.getSampleType())
+                "sample_type": api.get_title(obj.getSampleType()),
+                "default_retention_period": retention or "",
             }
 
     def get_reference_widget_attributes(self, name, obj=None):